chore: rimuove worktree gitlink, aggiunge .nojekyll, fix OutInWs retry
- Rimuove il gitlink fantasma sotto .worktrees (submodule senza
  .gitmodules) che faceva fallire il build di GitHub Pages.
- Aggiunge .nojekyll per servire i file statici senza Jekyll.
- OutInWs: retry limitato + null-check (vedi commit precedente).

+++ boundaryworker/utils/adapters/OutInWs.java
@@ -45,8 +45,7 @@ public class OutInWs implements IObserver{
     }
     
     private static final int MAX_RETRIES = 5;
-    private static final long RETRY_DELAY_MS = 2000;
-
+    private static final int RETRY_DELAY_MS = 2000;
     public void connectToRobotMind(String port) {
         for (int attempt = 1; attempt <= MAX_RETRIES; attempt++) {
             try {
